third_party/WebKit/LayoutTests/http/tests/css/resource-timing-details-for-revalidation-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS entry.domainLookupStart > 0 is true
This fetches ahem font.
